requirements.txt
opencv-python
face-recognition
numpy
pillow
requests
python-dotenv